+++ aderyn_core/src/ast/impls/node/events.rs
@@ -1,21 +1,7 @@
-use super::*;
+use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-
-#[derive(Clone, Debug, Deserialize, Eq, Serialize, PartialEq, Hash)]
-#[serde(rename_all = "camelCase")]
-pub struct EventDefinition {
-    pub anonymous: bool,
-    pub documentation: Option<Documentation>,
-    pub name: String,
-    pub name_location: Option<String>,
-    pub parameters: ParameterList,
-    pub event_selector: Option<String>,
-    pub src: String,
-    pub id: NodeID,
-}
 
 impl Node for EventDefinition {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
