+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -2,3 +2,9 @@ pub mod blocks;
 pub mod contracts;
 pub mod documentation;
 pub mod enumerations;
+pub mod errors;
+pub mod events;
+pub mod expressions;
+pub mod functions;
+pub mod identifiers;
+pub mod import_directives;
